Reorganize broker rake tests
Reorder tests in Rake test suite
