+++ broker/test/functional/rest_api_test.rb
@@ -1,9 +1,10 @@
+ENV["TEST_NAME"] = "functional_rest_api_test"
 require_relative '../test_helper'
 require 'openshift-origin-controller'
 require 'helpers/rest/api'
 require 'json'
 
-class RestApiUnitTest < ActionDispatch::IntegrationTest #ActiveSupport::TestCase
+class RestApiTest < ActiveSupport::TestCase
   def setup
     https!
     stubber
