+++ broker/test/functional_ext/rest_api_nolinks_test.rb
@@ -1,3 +1,4 @@
+ENV["TEST_NAME"] = "functional_ext_rest_api_nolinks_test"
 require 'test_helper'
 require 'helpers/rest/api_common'
 #set nolinks before loading rest api models
