+++ broker/test/unit/authentication_test.rb
@@ -1,3 +1,4 @@
+ENV["TEST_NAME"] = "unit_authentication_test"
 require File.expand_path('../../test_helper', __FILE__)
 
 class AuthenticationTest < ActiveSupport::TestCase
